+++ src/parser/inline/extension/label.ts
@@ -4,7 +4,7 @@ import { str } from '../../source';
 import { html } from 'typed-dom';
 import { join } from 'spica/array';
 
-const body = str(/^\$[a-z]*(?:(?:-[a-z][0-9a-z]*)+(?:-0(?:\.0){0,2}(?!\.?[A-Za-z0-9]))?(?!-?[A-Za-z0-9])|-(?:0|[1-9][0-9]*)(?:\.(?:0|[1-9][0-9]*)){0,2}(?![.-]?[A-Za-z0-9]))/);
+const body = str(/^\$[a-z]*(?:(?:-[a-z][0-9a-z]*)+(?!-?[A-Za-z0-9])|-(?:0|[1-9][0-9]*)(?:\.(?:0|[1-9][0-9]*)){0,2}(?![.-]?[A-Za-z0-9]))/);
 
 export const segment: ExtensionParser.LabelParser.SegmentParser = clear(validate(['[$', '$'], union([
   surround('[', body, ']'),
@@ -24,19 +24,11 @@ export const label: ExtensionParser.LabelParser = creator(fmap(
 export function number(label: string, base: string): string {
   return isFixed(label)
     ? label.slice(label.lastIndexOf('-') + 1)
-    : increment(
-        base,
-        isFormatted(label)
-          ? label.slice(label.lastIndexOf('-') + 1).split('.').length
-          : base.split('.').length);
+    : increment(base, base.split('.').length);
 }
 
 export function isFixed(label: string): boolean {
   return /^[^-]+-[0-9]+(?:\.[0-9]+)*$/.test(label);
-}
-
-export function isFormatted(label: string): boolean {
-  return /^[^-]+.+?-0(?:\.0)*$/.test(label);
 }
 
 function increment(number: string, position: number): string {